feat(*): 파일 확장자 확인하는 함수 함수로 분리
